requirements.txt
boto3
colorama
mysql-connector-python
python-dotenv
redis
ulid-py
requests